feat: thunk in playerDeck reducer
