+++ app/utils/player-utils.js
@@ -1,0 +1,28 @@
+import allCities from '../data/all-cities.js'
+import allEvents from '../data/all-events.js'
+import deckUtils from './deck-utils.js'
+
+export default {
+
+  initialDealPlayerCards: (numPlayers) => {
+    // pick playerhands and remove cards from deck
+    let numPlayerDrawnCards = 0
+    // game specifies specific num of cards per player depending on
+    // num of players
+    if (numPlayers === 2) numPlayerDrawnCards = 4
+    if (numPlayers === 3) numPlayerDrawnCards = 3
+    if (numPlayers === 4) numPlayerDrawnCards = 2
+    let playerHands = [] // all player hands
+    for (let i = 0; i < numPlayers; i++) {
+      let playerHand = [] // individual player hand
+      for (let j = 0; j < numPlayerDrawnCards; j++) {
+        playerHand.push(playerDeck.pop())
+      }
+      playerHands.push(playerHand)
+    }
+    // send playerhands to store & firebase
+    console.log("hands", playerHands)
+    initPlayerHands(playerHands)
+  }
+
+}
