+++ app/utils/playerDeck-utils.js
@@ -4,38 +4,21 @@ import deckUtils from './deck-utils.js'
 
 export default {
 
-  initializePlayerDeck: (numPlayers) => {
-    // randomize deck
-    // turn objects into arrays
+  // initial shuffle: returns all city & event cards shuffled together
+  initShufflePlayerDeck: (numPlayers) => {
+    // turn objects into arrays & shuffle
     const allCitiesArr = Object.keys(allCities).map((cityName) => {
       return { city: cityName, props: allCities[cityName] }
     })
-    // TODO: more efficient way to sort objects array?
     let playerDeck = allCitiesArr.concat(allEvents)
     deckUtils.shuffle(playerDeck)
-    // pick playerhands and remove cards from deck
-    let numPlayerDrawnCards = 0
-    // game specifies specific num of cards per player depending on
-    // num of players
-    if (numPlayers === 2) numPlayerDrawnCards = 4
-    if (numPlayers === 3) numPlayerDrawnCards = 3
-    if (numPlayers === 4) numPlayerDrawnCards = 2
-    let playerHands = [] // all player hands
-    for (let i = 0; i < numPlayers; i++) {
-      let playerHand = [] // individual player hand
-      for (let j = 0; j < numPlayerDrawnCards; j++) {
-        playerHand.push(playerDeck.pop())
-      }
-      playerHands.push(playerHand)
-    }
-    // send playerhands to store & firebase
-    console.log("hands", playerHands)
-    initPlayerHands(playerHands)
-    // randomize deck
-    // TODO: add epidemic cards
-    deckUtils.shuffle(playerDeck)
-    // send deck to firebase
     return playerDeck
   }
+
+      // pick playerhands and remove cards from deck
+
+    // TODO: add epidemic cards & shuffle
+    // deckUtils.shuffle(playerDeck)
+    // send deck to firebase
 
 }
